feat: move run query button back to the top of the page
