+++ packages/frontend/src/components/Explorer/ExplorerHeader/index.tsx
@@ -1,10 +1,13 @@
-import { Box, Group } from '@mantine/core';
+import { Badge, Box, Group, Tooltip } from '@mantine/core';
+import { IconAlertCircle } from '@tabler/icons-react';
 import { FC, memo, useEffect } from 'react';
 import useDashboardStorage from '../../../hooks/dashboard/useDashboardStorage';
 import { useExplorerContext } from '../../../providers/ExplorerProvider';
 import { Can } from '../../common/Authorization';
+import MantineIcon from '../../common/MantineIcon';
 import ShareShortLinkButton from '../../common/ShareShortLinkButton';
 import ExploreFromHereButton from '../../ExploreFromHereButton';
+import { RefreshButton } from '../../RefreshButton';
 import RefreshDbtButton from '../../RefreshDbtButton';
 import SaveChartButton from '../SaveChartButton';
 
@@ -17,6 +20,15 @@ const ExplorerHeader: FC = memo(() => {
     );
     const isValidQuery = useExplorerContext(
         (context) => context.state.isValidQuery,
+    );
+    const showLimitWarning = useExplorerContext(
+        (context) =>
+            context.queryResults.data &&
+            context.queryResults.data.rows.length >=
+                context.state.unsavedChartVersion.metricQuery.limit,
+    );
+    const limit = useExplorerContext(
+        (context) => context.state.unsavedChartVersion.metricQuery.limit,
     );
 
     const { getHasDashboardChanges } = useDashboardStorage();
@@ -43,6 +55,31 @@ const ExplorerHeader: FC = memo(() => {
                     <RefreshDbtButton />
                 </Box>
                 <Group spacing="xs">
+                    {showLimitWarning && (
+                        <Tooltip
+                            width={400}
+                            label={`Query limit of ${limit} reached. There may be additional results that have not been displayed. To see more, increase the query limit or try narrowing filters.`}
+                            multiline
+                            position={'bottom'}
+                        >
+                            <Badge
+                                leftSection={
+                                    <MantineIcon
+                                        icon={IconAlertCircle}
+                                        size={'sm'}
+                                    />
+                                }
+                                color="yellow"
+                                variant={'outline'}
+                                sx={{
+                                    textTransform: 'none',
+                                }}
+                            >
+                                Results may be incomplete
+                            </Badge>
+                        </Tooltip>
+                    )}
+                    <RefreshButton size={'xs'} />
                     {!savedChart && (
                         <Can I="manage" a="SavedChart">
                             <SaveChartButton isExplorer />
