+++ packages/frontend/src/components/Explorer/index.tsx
@@ -1,10 +1,12 @@
-import { Stack } from '@mantine/core';
+import { Alert, Stack, Text } from '@mantine/core';
 import { FC, memo } from 'react';
 import { useParams } from 'react-router-dom';
 
 import { ProjectType } from '@lightdash/common';
+import { IconAlertCircle } from '@tabler/icons-react';
 import { useProjects } from '../../hooks/useProjects';
 import { useExplorerContext } from '../../providers/ExplorerProvider';
+import MantineIcon from '../common/MantineIcon';
 import { CustomVisualizationProvider } from '../CustomVisualization';
 import DrillDownModal from '../MetricQueryData/DrillDownModal';
 import MetricQueryDataProvider from '../MetricQueryData/MetricQueryDataProvider';
@@ -25,6 +27,15 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
         const unsavedChartVersionMetricQuery = useExplorerContext(
             (context) => context.state.unsavedChartVersion.metricQuery,
         );
+        const showLimitWarning = useExplorerContext(
+            (context) =>
+                context.queryResults.data &&
+                context.queryResults.data.rows.length >=
+                    context.state.unsavedChartVersion.metricQuery.limit,
+        );
+        const limit = useExplorerContext(
+            (context) => context.state.unsavedChartVersion.metricQuery.limit,
+        );
         const { projectUuid } = useParams<{ projectUuid: string }>();
 
         const { data: projects } = useProjects({ refetchOnMount: false });
@@ -41,6 +52,22 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
             >
                 <Stack sx={{ flexGrow: 1 }}>
                     {!hideHeader && <ExplorerHeader />}
+
+                    {showLimitWarning && (
+                        <Alert
+                            icon={<MantineIcon icon={IconAlertCircle} />}
+                            color="yellow"
+                            title="Results may be incomplete"
+                            variant={'outline'}
+                        >
+                            <Text color="gray.6">
+                                Query limit of {limit} reached. There may be
+                                additional results that have not been displayed.
+                                To see more, increase the query limit or try
+                                narrowing filters.
+                            </Text>
+                        </Alert>
+                    )}
 
                     <FiltersCard />
